requirements.txt
fastapi
uvicorn
python-multipart
aiofiles

transformers
torch
sentencepiece
nltk
whisper
openai-whisper

moviepy
pydub
ffmpeg-python

easyocr
opencv-python

google-api-python-client
google-auth
google-auth-oauthlib
google-cloud-aiplatform

requests
shutil
pillow
tqdm

gradio
python-dotenv
